requirements.txt
python-jobspy>=1.1.80
python-dotenv>=1.0.0
pdfminer.six>=20221105
python-docx>=1.1.0
rich>=13.7.0
click>=8.1.7
SQLAlchemy>=2.0.0
tabulate>=0.9.0
flask>=3.0.0
werkzeug>=3.0.0
